added handling so variablelist could set set ccw32=ccw32 from and input like ccw32 This handles items normally in the path, that may need an absolute path
converted commonmenu path to variable and set it to menus in vimod.variables
added preset variable called newfile to set if want to clobber old file with append to file.

Modified Getline to handle getting first line as well
changed counted menu writeline to writeitem so it gets list as before. It is only on matchin that the line if getlined